requirements.txt
Flask==3.0.2
openai==1.41.1
python-dotenv==0.19.2
gunicorn==20.0.4
Werkzeug==3.0.1